Close the crash and leak batch from the sibling port's stabilisation pass
Ported by meaning from DeadAir-Engine-x64-OpenSource 2c50641, ce719e1 and
df677f9, each hunk checked against our tree first; what we already had, or had
fixed another way, was left alone. Everything below was live in ours.

Crashes, most reachable from scripts or save data:
- level.vertex_id summed u32(-1)+1 in 32 bits and handed Lua vertex 0 for
  "no vertex" - a real node every script-side guard lets through, and a
  permanent re-path storm for whoever trusted it. The sum is 64-bit now.
- CUICellContainer::PickCell divided by a container slot capacity of 0 from
  config - C0000094 on dragging an item over it.
- SGameTaskObjective::CreateMapLocation: on load the map spot is searched,
  not created; a miss left a null pointer behind a VERIFY and complex_spot()
  read it. The objective now goes unmarked, and says so.
- CGameObject::update_ai_locations: three VERIFYs gone in release and an
  inverted "!graph && table" condition; a search miss (u32(-1)) went times six
  into the cross table. Live checks, the right condition, and the cross table's
  own bound.
- CALifeSpawnRegistry::load_updates: the vertex id comes from the save; a
  save made on another spawn dereferenced null on load. Such updates are
  skipped with a message, the rest of the save loads.
- CScriptParticles: the engine frees a finished effect and nulls the pointer,
  every later Lua call was a null dereference (StartPath/StopPath/PausePath
  had no check at all).
- CHARACTER_COMMUNITY::relation/set_relation/sympathy: an unknown faction
  name from Lua is index -1, read and written.
- CObjectList relcase walk: a callback that unregisters itself swaps-with-back
  under the iterator; the walk is by index now, and relcase_unregister refuses
  a stale id instead of writing past the vector.
- CSE_ALifeOnlineOfflineGroup: a member pointer may name a released object;
  unregister_member, synchronize_location and switch_offline now ask the live
  object registry (the same one the script accessors use) and take the first
  live member.
- Bloodsucker alien/vampire: Actor() dereferenced on deactivate and cleanup,
  which run from the monster's own destruction where the actor may be gone.
- doors: a stalker reverting its doors from its destructor handed Lua its
  half-destroyed game object through the use callback (read at -1). The door
  still changes state; only the notification is skipped.
- uber_deffer: an empty detail-bump name passed the pointer test (c_str() of
  an empty shared_str is "", not null), leaving a texture named "#".

Leaks and ratchets:
- GameEventQueue and INetQueue pools grew to the all-time peak at 16 KB a
  cell and never shrank - the minute-of-quiet pruning does not fire in play.
  Capped at a 256-cell reserve.
- LogFile kept every line of the session in memory; trimmed to the last
  40000, which is more than the console and the crash report ever read.

Cosmetic, but reported:
- Condition bars on pooled cell widgets: one eatable switched the gradient
  off for every bar drawn after it.
- An empty slot drew its condition track at zero; visibility is driven now.
- Space/Enter confirm the level-change prompt, gated on the allowed case.

+++ src/xrGame/level_script.cpp
@@ -928,9 +928,11 @@ void CLevel::script_register(lua_State* luaState)
 
         def("vertex_id", +[](Fvector position) -> u64
         {
-            // Original luabind converts 4294967295 (which is u32(-1)) to 4294967296
+            // Original luabind converts 4294967295 (which is u32(-1)) to 4294967296. The sum has
+            // to be taken in 64 bits: in u32 it wraps to 0, and "no vertex" reaches Lua as vertex 0
+            // - a real node in a corner of the map that every script-side guard lets through.
             const u32 id = ai().level_graph().vertex_id(position);
-            return id == u32(-1) ? id + 1 : id; // reproduce original behaviour
+            return id == u32(-1) ? u64(id) + 1 : u64(id); // reproduce original behaviour
         }),
         def("game_id", &GameID),
         def("ray_pick", &ray_pick)
